requirements.txt
# Core Dependencies
streamlit>=1.28.0
openai>=1.0.0
pandas>=1.5.0
numpy>=1.24.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Document Processing
openpyxl>=3.1.0
python-docx>=0.8.11
python-pptx>=0.6.21
pdfplumber>=0.9.0
pytesseract>=0.3.10

# Natural Language Processing
nltk>=3.8.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=1.7.0
joblib>=1.3.0

# Google Drive API
google-api-python-client>=2.100.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# Date Utilities
python-dateutil>=2.8.0
